# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mstore API Documentation",
        "description": "API documentation for Mstore E-commerce Backend",
        "contact": {
            "email": "support@mstore.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/currency": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Get list of currencies",
                "description": "Returns all available currencies for the store.",
                "operationId": "getCurrencies",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "field",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "USD"
                                                    },
                                                    "symbol": {
                                                        "type": "string",
                                                        "example": "$"
                                                    },
                                                    "no_of_decimal": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "exchange_rate": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "status": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_by_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Create a new currency",
                "description": "Add a new currency with exchange rate.",
                "operationId": "createCurrency",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "symbol",
                                    "exchange_rate",
                                    "status"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "ISO 4217 currency code",
                                        "type": "string",
                                        "maxLength": 3,
                                        "example": "EUR"
                                    },
                                    "symbol": {
                                        "description": "Currency symbol",
                                        "type": "string",
                                        "example": "€"
                                    },
                                    "no_of_decimal": {
                                        "description": "Decimal places",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "exchange_rate": {
                                        "description": "Exchange rate from base currency",
                                        "type": "number",
                                        "example": 0.92
                                    },
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Currency created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/currency/{id}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Get currency by ID",
                "operationId": "getCurrencyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Currency not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Update currency",
                "operationId": "updateCurrency",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "symbol": {
                                        "type": "string"
                                    },
                                    "exchange_rate": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currency updated"
                    },
                    "404": {
                        "description": "Currency not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Delete currency",
                "operationId": "deleteCurrency",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency deleted"
                    },
                    "404": {
                        "description": "Currency not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/currency/{id}/{status}": {
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Update currency status",
                "operationId": "updateCurrencyStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/currency/deleteAll": {
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Delete multiple currencies",
                "operationId": "deleteMultipleCurrencies",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currencies deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}